Structured the app routes
